Add tags to urls for page hit tracking
Courtesy of Paul, we have a tracking code for our tutorial. I appended a tag to the end of all links to Microsoft documentation pages or properties like Azure/VS Code/PowerBI. 'msgarage' is our unique tracking id that will allow them to aggregate the results into a dashboard for measuring impact.